Use multi-stage build to reduce image size
Using the new multi-stage build feature in Docker 17.05, we can
reduce the size of this image by ~50%. Only ffmpeg is copied
and only the necessary dependencies are installed in the
resulting runtime image.

This also has the additional benefit of making it clear what the
runtime dependencies are in the Dockerfile.